backend/requirements.txt
Django>=5.2,<6
djangorestframework>=3.15,<4
djangorestframework-simplejwt>=5.3,<6
django-cors-headers>=4.6,<5
Pillow>=11,<12
